adjust how anchors are created

+++ include/kbbq/readutils.hh
@@ -58,8 +58,8 @@ namespace readutils{
 			std::vector<bool> not_skipped_errors() const;
 			//fill errors attribute given the bloom filter and thresholds.
 			void infer_read_errors(const bloom::bloomary_t& b, const std::vector<int>& thresholds, int k);
-			//fix one error and return the length of kmers it fixed
-			int correct_one(const bloom::bloomary_t& t, int k);
+			//fix one error and return the index of the fixed base; std::string::npos if no fixes are found
+			size_t correct_one(const bloom::bloomary_t& t, int k);
 			static void load_rgs_from_bamfile(bam_hdr_t* header);
 			void get_errors(const bloom::bloomary_t& trusted, int k, int minqual = 6);
 			std::vector<int> recalibrate(const covariateutils::dq_t& dqs, int minqual = 6) const;
